moved Header and NavigationBar
