+++ src/layouts/DashboardLayout.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 
-import Header from '../components/generic/Header'
-import NavigationBar from '../components/generic/NavigationBar'
+import Header from '../components/common/Header'
+import NavigationBar from '../components/common/NavigationBar'
 
 const DashboardLayout = ({ children }) => {
   const [sidebarOpen, setSidebarOpen] = useState(false)
